ai-backend-ergo-app-prod/streamlit_app/requirements.txt
streamlit>=1.29.0
anthropic>=0.18.1
Pillow>=10.1.0
python-dotenv>=1.0.0
